Added javafx library/modules and skeleton for the ui
