+++ src/main/java/views/ChatScreen.java
@@ -1,0 +1,2 @@
+package views;public class ChatScreen {
+}
