+++ src/main/java/views/ChatScreenController.java
@@ -1,0 +1,2 @@
+package views;public class ChatScreenController {
+}
