+++ src/main/java/views/ChatScreenPresenter.java
@@ -1,0 +1,2 @@
+package views;public class ChatScreenPresenter {
+}
